requirements.txt
openai>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
reportlab>=4.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
google-search-results>=2.4.2
fastapi>=0.100.0
uvicorn>=0.23.0
sse-starlette>=1.8.0
aiofiles>=23.2.0
